Flask-App/requirements.txt
absl-py==1.3.0
alembic==1.9.0
appdirs==1.4.4
asttokens==2.0.8
attrs==22.1.0
autopage==0.5.1
backcall==0.2.0
blis==0.7.9
cachetools==5.2.0
catalogue==2.0.8
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
cliff==4.1.0
cmaes==0.9.0
cmd2==2.4.2
colorama==0.4.5
colorlog==6.7.0
confection==0.0.3
contourpy==1.0.5
cssselect2==0.7.0
cv==1.0.0
cycler==0.11.0
cymem==2.0.7
Cython==0.29.32
debugpy==1.6.3
decorator==5.1.1
entrypoints==0.4
et-xmlfile==1.1.0
executing==1.1.0
fastjsonschema==2.16.2
Flask==2.2.2
fonttools==4.37.3
fst-pso==1.8.1
FuzzyTM==2.0.5
google-auth==2.14.1
google-auth-oauthlib==0.4.6
greenlet==2.0.1
grpcio==1.50.0
icecream==2.1.3
idna==3.4
imageio==2.22.4
importlib-metadata==4.13.0
imutils==0.5.4
ipykernel==6.16.0
ipynb==0.5.1
ipyplot==1.1.1
ipython==8.5.0
itsdangerous==2.1.2
jedi==0.18.1
Jinja2==3.1.2
joblib==1.2.0
jsonschema==4.16.0
jupyter-core==4.11.1
jupyter_client==7.3.5
kiwisolver==1.4.4
langcodes==3.3.0
lime==0.2.0.1
lxml==4.9.2
Mako==1.2.4
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.6.0
matplotlib-inline==0.1.6
matplotlib-scalebar==0.7.2
miniful==0.0.6
murmurhash==1.0.9
nbformat==5.7.0
nest-asyncio==1.5.5
networkx==2.8.8
nltk==3.8.1
numpy==1.23.3
oauthlib==3.2.2
opencv-python==4.6.0.66
openpyxl==3.0.10
optuna==3.0.5
packaging==21.3
palettable==3.3.0
pandas==1.5.0
parso==0.8.3
pathy==0.10.1
pbr==5.11.0
pi==0.1.2
pickleshare==0.7.5
Pillow==9.2.0
plotly==5.10.0
pooch==1.6.0
preshed==3.0.8
prettytable==3.5.0
prompt-toolkit==3.0.31
protobuf==3.20.3
psutil==5.9.2
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydantic==1.10.4
pyFUME==0.2.25
Pygments==2.13.0
pyparsing==3.0.9
pyperclip==1.8.2
pyreadline3==3.4.1
pyrsistent==0.18.1
python-dateutil==2.8.2
pytz==2022.2.1
PyWavelets==1.4.1
PyYAML==6.0
pyzmq==24.0.1
regex==2022.10.31
reportlab==3.6.12
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
scikit-image==0.19.3
scikit-learn==1.1.2
scipy==1.8.1
seaborn==0.12.0
seaborn-image==0.5.0
shortuuid==1.0.11
simpful==2.9.0
six==1.16.0
sklearn==0.0

smart-open==6.3.0
SQLAlchemy==1.4.45
srsly==2.4.5
stack-data==0.5.1
stevedore==4.1.1
svglib==1.4.1
tabulate==0.9.0
tenacity==8.1.0

thinc==8.1.6
threadpoolctl==3.1.0
tifffile==2022.10.10
tinycss2==1.2.1

tornado==6.2
tqdm==4.64.1
traitlets==5.4.0
typer==0.7.0
typing_extensions==4.4.0
urllib3==1.26.12
wasabi==0.10.1
wcwidth==0.2.5
webencodings==0.5.1

numpy
pandas
Flask>=2.2.2
Jinja2>=3.1.2
MarkupSafe==2.1.1
Werkzeug==2.2.2
torch==1.13.0 -f https://download.pytorch.org/whl/cpu
torchvision==0.14.0 -f https://download.pytorch.org/whl/cpu
torchmetrics==0.10.2
imageio==2.22.4
opencv-python==4.6.0.66
requests==2.28.1
requests-oauthlib==1.3.1
zipp==3.11.0
webencodings==0.5.1
urllib3==1.26.12
pickleshare==0.7.5
Pillow==9.2.0
plotly==5.10.0
matplotlib==3.6.0
matplotlib-inline==0.1.6
matplotlib-scalebar==0.7.2
PyYAML==6.0
svglib==1.4.1